Add: automated build & itch upload workflow
Note: current implementation happens on all branch pushes, not ideal.
Should adjust this before merging

Implementation relied on examples from a few sources:
* https://github.com/abarichello/godot-ci/blob/master/.github/workflows/godot-ci.yml#L63
* https://github.com/Oval-Tutu/publish-to-itch-with-butler/blob/main/action.yml

Oval-Tutu's action ended up having some buggy behavior, so I
re-implemented the same general logic myself. May be a good idea to
offer a contribution in the future to improve it for others in the
future.